+++ src/main/java/com/example/dal/CustomerDAO.java
@@ -7,30 +7,9 @@ import org.springframework.stereotype.Component;
 
 import java.util.List;
 
-@Component
-public class CustomerDAO {
 
-    private JdbcTemplate jdbcTemplate;
+public interface CustomerDAO {
 
-    @Autowired
-    public CustomerDAO(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public List<Customer> findCustomersByName(String name) {
-        return jdbcTemplate.query(
-                "SELECT id, first_name, last_name FROM customers WHERE first_name = ?",
-                new Object[] { name },
-                (rs, rowNum) ->
-                        new Customer(rs.getLong("id"),
-                                rs.getString("first_name"),
-                                rs.getString("last_name"))
-        );
-    }
-
-    public void createCustomers(List<Object[]> splitUpNames) {
-        //splitUpNames is list of object[] where [0] is first name and [1] is the last name
-        // Uses JdbcTemplate's batchUpdate operation to bulk load data
-        jdbcTemplate.batchUpdate("INSERT INTO customers(first_name, last_name) VALUES (?,?)", splitUpNames);
-    }
+    List<Customer> findCustomersByName(String name);
+    void createCustomers(List<Object[]> splitUpNames);
 }
