+++ src/main/java/com/example/service/FileProcessorService.java
@@ -1,0 +1,10 @@
+package com.example.service;
+
+import java.util.List;
+
+
+public interface FileProcessorService {
+
+    List<String> getFileContents(String filename);
+
+}
