+++ src/test/resources/test1.txt
@@ -1,0 +1,4 @@
+this
+is
+a
+test
