+++ test2.txt
@@ -1,0 +1,3 @@
+test
+2
+stinks
